feat: дбавил валидацию по сценариям, добавил константу заголовка, добавил интерфейсы сервисов.
