+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -3,10 +3,11 @@ package ru.practicum.shareit.user.dto;
 import lombok.Builder;
 import lombok.Data;
 import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.validation.Create;
+import ru.practicum.shareit.validation.Update;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 /**
  * DTO-класс пользователя, копирует поля {@link User}.
@@ -15,10 +16,9 @@ import javax.validation.constraints.NotNull;
 @Builder(toBuilder = true)
 public class UserDto {
     private int id;
-    @NotNull(message = "Имя или логин не может отсутствовать.")
-    @NotBlank(message = "Имя или логин не может быть пустым или содержать пробелы.")
+    @NotBlank(message = "Имя или логин не может быть пустым или содержать пробелы.", groups = {Create.class})
     private String name;
-    @NotNull(message = "Электронная почта не может отсутствовать.")
-    @Email(message = "Некорректно указана электронная почта.")
+    @NotBlank(message = "Электронная почта не может быть пустой или содержать пробелы.", groups = {Create.class})
+    @Email(message = "Некорректно указана электронная почта.", groups = {Update.class, Create.class})
     private String email;
 }
