+++ src/main/java/ru/practicum/shareit/validation/Create.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.validation;
+
+/**
+ * Интерфейс-указатель группы валидации.
+ */
+public interface Create {
+}
